Engine Node + Cluster Test Util Updates (#46)
* update authority fixture

* added documentation for test utils in narwhal-types

* move network error to network crate

* refactor test utils for better organization

* refactor and remove unused imports

* fix

* cargo fmt

* init outline for cluster to spawn execution components

* refactor builder and execution tests - pass args to methods instead of redundantly parsing NodeCommand

* unit test for adjusting rpc ports

* update unit test to ensure 4 node clusters are updated correctly. probably need to add network arg adjustment too for discovery port

* change default network disc port per instance as well

* layout for worker's batch maker

* organizational notes

* remove narwhal-node package - doesn't work with rust analyzer and is sloppy

* tn-node uses error types instead of unwrap

* refactor start / spawn sequence in primary & worker nodes

* started moving execution types in tn-node start sequence

* remove <State> generic from tn-node and cluster methods

* working through exection node impl - need to separate worker / primary instances instead of just one each - but maybe I'll skip for now since only one worker per primary for now

* stuck on generic mismatch

* .

* abandon generics approach for now - compiliation checkpoint

* compilation checkpoint - abandon generics all together D:

* rough outline for starting el executor for primary

* .

* worker batch maker started

* spawn el executor within start_primary method for cluster

* fmt

* fix primary tests

* comment out consensus integration test for cargo fix

* compilation for worker tests

* remove useless worker test

* make Cluster::new async to grab current handle for runtime

* adjust ports on inner type for execution node

* basic cluster test passing, but I think the execution node portion is failing - beacon consensus engine is immediately unavailable when sending forkchoice update inside inner::start_engine - integration test works, so there's something else happening with tasks spawning

* add beacon forkchoice update failure - consensus test fails

* .

* hold task manager, not executor for engine components. beacon engine spawning now

* concept for shutting down engine / workers based on separate task managers

* remove async from cluster new / engine new - runtime handle only needed in async  methods

* shutdown unit test for engine

* atomic bools updated when engine/worker tasks start/stop. for some reason, the worker channel never gets dropped. moving on for now

* shutdown all method - updates to cluster's authority details to call execution node

* consider spawn critical - cluster test failing now w/ error in rx_notifier dropping

* change worker id to u16

* batch maker handling multiple workers

* refactor workers in execution node

* fmt - spawn executor from manager at time of creation

* graceful shutdown test doesn't work very well - seems to be overkill - moving on

* refactor shutdown - process working now - network port collision happening - need to adjust ports further

* executor engine test passing

* leaving engine test as is for now

* cargo fmt

* new batch validator

* refactor batch validator with more power

* resort to explicit types for new batch validator method

* create keypair for authority EL

* add execution layer info to authority struct - rename primary address to reference network

* update tests for primary_network_address method

* add address to execution node

* fmt

* use batch validation trait for spawning worker

* test utils for noop batch validation and nil batch validation

* add beneficiary address to batches

* batch maker test for execution node - added rpc http local addr method to inner node

* return http client for rpc server

* batch maker test for execution node

* extend batch maker test to assert rpc is shutdown with worker

* breaking commit - need to update test for network args on workers

* cluster test working now

* cleanup

* cargo fmt

* update engine node test for port creation

* more cleanup - ready for PR

* assert first 3 blocks are equal for each authority in basic cluster test

* clear warnings for more tests

+++ crates/consensus/types/src/config/mod.rs
@@ -29,8 +29,10 @@ use utils::get_available_port;
 /// The epoch number.
 pub type Epoch = u64;
 
-// Opaque bytes uniquely identifying the current chain. Analogue of the
-// type in `sui-types` crate.
+/// Opaque bytes uniquely identifying the current chain. Analogue of the
+/// type in `sui-types` crate.
+///
+/// TODO: this should be replaced with chainspec.
 #[derive(Clone, Serialize, Deserialize, Debug, Eq, PartialEq)]
 pub struct ChainIdentifier([u8; 32]);
 
@@ -115,7 +117,7 @@ impl<S: Serialize> Export for S {}
 // TODO: This actually represents voting power (out of 10,000) and not amount staked.
 // Consider renaming to `VotingPower`.
 pub type Stake = u64;
-pub type WorkerId = u32;
+pub type WorkerId = u16;
 
 /// Holds all the node properties. An example is provided to
 /// showcase the usage and deserialization from a json file.
@@ -362,6 +364,10 @@ impl Parameters {
     }
 }
 
+/// Information for the worker
+/// - [NetworkPublicKey] for identifying network messages.
+/// - [Multiaddr] for receiving transactions.
+/// - [Multiaddr] for receiving messages from other workers and the primary.
 #[derive(Clone, Serialize, Deserialize, Eq, Hash, PartialEq, Debug)]
 pub struct WorkerInfo {
     /// The public key of this worker.
@@ -372,9 +378,14 @@ pub struct WorkerInfo {
     pub worker_address: Multiaddr,
 }
 
+/// Map of all workers for the authority.
+///
+/// The map associates the worker's id to [WorkerInfo].
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct WorkerIndex(pub BTreeMap<WorkerId, WorkerInfo>);
 
+/// The collection of all workers organized by authority public keys
+/// that comprise the [Committee] for a specific [Epoch].
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct WorkerCache {
     /// The authority to worker index.
